# Conflicts:
#   frontend/src/pages/MainPage.tsx

Former-commit-id: 73148e58f484be1fea8b6c9f8d39c156481146c8

+++ backend/src/main/java/com/ssafy/api/service/FriendshipRepository.java
@@ -9,15 +9,15 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.Friendship;
 
 @Repository
-public interface FriendshipRepository extends JpaRepository<Friendship, Integer>{
+public interface FriendshipRepository extends JpaRepository<Friendship, Long>{
 
-	Optional<Friendship> findByFriend1IDAndFriend2ID(int user1id, int user2id);
+	Optional<Friendship> findByFriend1IDAndFriend2ID(long user1id, long user2id);
 
-	List<Friendship> findByFriend1ID(int userID);
+	List<Friendship> findByFriend1ID(long userID);
 
-	List<Friendship> findByFriend2ID(int userID);
+	List<Friendship> findByFriend2ID(long userID);
 
-	void deleteByFriend1ID(int userID);
+	void deleteByFriend1ID(long userID);
 
-	void deleteByFriend2ID(int userID);
+	void deleteByFriend2ID(long userID);
 }
